Fix StageState not being created when AppUsage is disabled
